Register functionality added

+++ base/views.py
@@ -1,33 +1,63 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 from django.contrib.auth import authenticate, login, logout
 from django.contrib import messages
-
+from django.contrib.auth.forms import UserCreationForm
 from .models import Room, Topic
 from .forms import RoomForm
 
 
 def loginPage(request):
+    page = 'Login'
+    if request.user.is_authenticated:
+        return redirect('home')
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
         try:
             user = User.objects.get(username=username)
             user = authenticate(request, username=username, password=password)
             if user is not None:
                 login(request, user)
+                messages.success(request, 'Hello ' + user.get_username())
                 return redirect('home')
             else:
                 messages.error(request, 'Username or password is incorrect')
         except:
             messages.error(request, 'User does not exist')
 
-    return render(request, 'register_login.html')
+    context = {'page': page}
+    return render(request, 'register_login.html', context)
+
+
+def registerPage(request):
+    page = 'Register'
+    form = UserCreationForm()
+    if request.user.is_authenticated:
+        return redirect('home')
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            messages.success(request, 'Hello ' + user.get_username())
+            return redirect('home')
+        else:
+            messages.error(request, 'Something went wrong!')
+    
+    context = {'page': page, 'form': form}
+    return render(request, 'register_login.html', context)
+
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
 
 def home(request):
     query = request.GET.get('q')
@@ -39,12 +69,15 @@ def home(request):
     context = {'rooms': rooms, 'topics': topics, 'rooms_count':rooms.count()}
     return render(request, 'home.html', context)
 
+
 def room(request, pk):
     room = Room.objects.get(id=pk)
    
     context = {'room': room}
     return render(request, 'room.html', context)
 
+
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -55,10 +88,15 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'room_form.html', context)
 
+
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
     
+    if request.user != room.host.username:
+        return HttpResponse('Access Denied!')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -68,9 +106,14 @@ def updateRoom(request, pk):
     context = { 'form' : form }
     return render(request, 'room_form.html', context)
 
+
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
-    print('delete route')
+
+    if request.user != room.host.username:
+        return HttpResponse('Access Denied!')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
